% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generationFunctions.R
\name{compareTrajectoryData}
\alias{compareTrajectoryData}
\title{Compare two trajectory dataframes}
\usage{
compareTrajectoryData(observedData, generatedData)
}
\arguments{
\item{observedData}{A data.frame object which is output of Cohort2Trajectory package}

\item{generatedData}{The data frame of generated data}
}
\description{
This function outputs a data table comparing two patient trajectory dataframes
}
\keyword{internal}
