% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generationFunctions.R
\name{compareTrajectoryDataLogRank}
\alias{compareTrajectoryDataLogRank}
\title{Compare two trajectory dataframes with logRank tests for all possible transfers}
\usage{
compareTrajectoryDataLogRank(observedData, generatedData)
}
\arguments{
\item{generatedData}{the data frame of generated data}

\item{observedDataA}{data.frame object which is output of Cohort2Trajectory package}
}
\description{
This function outputs a matrix table comparing two patient trajectory dataframes with logRank tests
}
