% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpTableFunctions.R
\name{createIndex}
\alias{createIndex}
\title{Function for creating an index non-clustered/clustered, unique/non-unique}
\usage{
createIndex(
  connection,
  dbms = "postgresql",
  tableSchema = "",
  relationName,
  columnNames,
  label,
  clustered = FALSE,
  unique = FALSE,
  tempTable = FALSE
)
}
\arguments{
\item{connection}{Connection to the database (package DatabaseConnector)}

\item{dbms}{Database dialect}

\item{tableSchema}{Schema in which the targeted table resides}

\item{relationName}{Name of the targeted table}

\item{columnNames}{Columns on which to build the index on}

\item{label}{Name for the index}

\item{clustered}{Boolean for clustered index}

\item{tempTable}{Boolean indicating whether the table in a "TEMP" table, if so, no schema is specified}
}
\description{
Function for creating an index non-clustered/clustered, unique/non-unique
}
\keyword{internal}
