% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcFunctions.R
\name{dataPreparation}
\alias{dataPreparation}
\title{Prepare patientData dataframe for msm package}
\usage{
dataPreparation(patientData)
}
\arguments{
\item{cohortData}{A data.frame object which is output of Cohort2Trajectory package}
}
\value{
A dataframe ready for using msm package
}
\description{
This function prepares patientData dataframe for msm package
}
\keyword{internal}
