% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{daysOverlap}
\alias{daysOverlap}
\title{Function for calculating overlap in date intervals}
\usage{
daysOverlap(dateStart, dateEnd, controlStart, controlEnd)
}
\arguments{
\item{dateStart}{Start of the boundary date interval}

\item{dateEnd}{End of the boundary date interval}

\item{controlStart}{Start of the date interval of interest}

\item{controlEnd}{End of the date interval of interest}
}
\description{
Function for calculating overlap in date intervals
}
\keyword{internal}
