% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{drawSunburst}
\alias{drawSunburst}
\title{Visualise patients' trajectories with sunburst plot}
\usage{
drawSunburst(patientData)
}
\arguments{
\item{pateintData}{A data.frame object which is output of Cohort2Trajectory package data.frame object which is resulting from}
}
\description{
Visualise patients' trajectories with sunburst plot
}
\keyword{internal}
