% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{droppingTables}
\alias{droppingTables}
\title{Function for asking from user whether to drop the package-generated tables from database}
\usage{
droppingTables()
}
\description{
Function for asking from user whether to drop the package-generated tables from database
}
\keyword{internal}
