% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generationFunctions.R
\name{generateDataDiscrete}
\alias{generateDataDiscrete}
\title{Generate patient trajectories}
\usage{
generateDataDiscrete(
  transitionMatrix,
  n = 100,
  minDate = "1900-01-01",
  maxDate = "2021-12-31",
  maxOut = 183,
  stateDuration = 30,
  pathToResults = getwd(),
  generateCost = 0,
  statisticsTable = NULL,
  studyName = ""
)
}
\arguments{
\item{transitionMatrix}{The transition matrix for states}

\item{n}{Number of patients' trajectories to generate}

\item{minDate}{Smallest possible trajectory start date}

\item{maxDate}{Largest possible trajectory start date}

\item{maxOut}{Maximum nr of days patient can be out of cohort}

\item{pathToResults}{Path to temp folder for saving objects}

\item{generateCost}{Value [0,1] indicating whether to generate cost of states}

\item{statisticsTable}{Table retrieved from observed data for cost values}

\item{studyName}{Customized study name}
}
\description{
This function outputs a data.frame object which describes the movement of patients between the defined states and state duration
}
