% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generationFunctions.R
\name{generateStateCost}
\alias{generateStateCost}
\title{Compare two trajectory dataframes with logRank tests for all possible transfers}
\usage{
generateStateCost(observedStatistics, state, duration)
}
\arguments{
\item{observedStatistics}{Statistics observed from data (mean, std) dataframe}

\item{state}{The state notation}

\item{duration}{Duration of the state}
}
\description{
This function outputs a value generated from observed data
}
\keyword{internal}
