% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpTableFunctions.R
\name{generateTempTables}
\alias{generateTempTables}
\title{Create temp tables for package}
\usage{
generateTempTables(
  connection,
  dbms,
  cdmSchema = "ohdsi_cdm",
  cdmTmpSchema = "ohdsi_temp"
)
}
\arguments{
\item{connection}{Connection to database}

\item{dbms}{Database management system: postgresql, mysql etc.}

\item{cdmTmpSchema}{tmpSchema}
}
\value{
Boolean of success
}
\description{
This function creates all temp tables for running TrajectoryMarkovAnalysis package
}
\keyword{internal}
