% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{getCost}
\alias{getCost}
\title{Get patient's drug era cost}
\usage{
getCost(
  connection,
  dbms,
  patientId,
  startDate,
  endDate,
  domain = c("Drug", "Visit", "Procedure", "Device", "Measurement", "Observation",
    "Specimen"),
  cdmTmpSchema
)
}
\arguments{
\item{connection}{Connection to database}

\item{dbms}{Database management system: postgresql, mysql etc.}

\item{patientId}{The id number of the patient of interest}

\item{startDate}{Starting date of the observation}

\item{endDate}{Ending date of the observation}

\item{domain}{Vector of treatment types of interest (Drug, Visit, Procedure, Device, Measurement, Observation, Specimen)}

\item{cdmTmpSchema}{temp tables schema}
}
\value{
Value of complete cost
}
\description{
This function queries patient's treatment cost
}
\keyword{internal}
