% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{getCostDistPlot}
\alias{getCostDistPlot}
\title{Get patients' state statistics}
\usage{
getCostDistPlot(connection, dbms, cdmTmpSchema)
}
\arguments{
\item{connection}{Connection to database}

\item{dbms}{The database dialect}

\item{cdmTmpSchema}{Temp tables' schema}
}
\value{
plotplot
}
\description{
Get patients' state statistics
}
\keyword{internal}
