% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionTreeFunctions.R
\name{getDTreeProbabilities}
\alias{getDTreeProbabilities}
\title{Get a dataframe of patients' trajectories}
\usage{
getDTreeProbabilities(cohortData, stateCohorts, pathToResults = getwd())
}
\arguments{
\item{cohortData}{A data.frame object which is output of Cohort2Trajectory package}

\item{stateCohorts}{The IDs of state cohorts}

\item{pathToResults}{Path to temp folder for saving objects}
}
\description{
This function outputs a data.frame object which describes the movement of patients between the defined states and state duration
}
\keyword{internal}
