% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{getFirstState}
\alias{getFirstState}
\title{Get patients' first state in the main cohort and their inclusion time}
\usage{
getFirstState(cohortData, excludedStates = c())
}
\arguments{
\item{cohortData}{A data.frame object which is output of Cohort2Trajectory package}

\item{excludedStates}{A vector populated with excluded states}
}
\value{
data.frame with four columns: unique SUBJECT_ID, FIRST_STATE, SUBJECT_START_DATE, SUBJECT_END_DATE
}
\description{
Get patients' first state in the main cohort and their inclusion time
}
\keyword{internal}
