% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{getStateStatistics}
\alias{getStateStatistics}
\title{Get patients' state and trajectory cost statistics}
\usage{
getStateStatistics(
  connection,
  dbms,
  cohortData,
  cdmTmpSchema,
  stateStatistics = NULL,
  studyName = "THE-STUDY",
  excludedStates = c(),
  cost_domains = c("Drug", "Visit", "Procedure", "Device", "Measurement", "Observation",
    "Specimen")
)
}
\arguments{
\item{connection}{Connection to the database}

\item{dbms}{Database dialect}

\item{cohortData}{A data.frame object which is output of Cohort2Trajectory package}

\item{cdmTmpSchema}{The temp tables' schema}

\item{stateStatistics}{A (reactive) table for the retention of state statistics}

\item{excludedStates}{A vector populated with excluded states}

\item{cost_domains}{The selected cost domain ids}
}
\value{
A data.frame with each state's statistics
}
\description{
Get patients' state and trajectory cost statistics
}
\keyword{internal}
