% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryFunctions.R
\name{getStohasticMatrix}
\alias{getStohasticMatrix}
\title{Get Markov state probabilities from raw data with defined state duration}
\usage{
getStohasticMatrix(
  cohortData,
  stateCohorts,
  pathToResults = getwd(),
  statePriorityVector = NULL,
  studyName = "THE-STUDY",
  excludedStates = NULL
)
}
\arguments{
\item{cohortData}{A data.frame object which is output of Cohort2Trajectory package}

\item{stateCohorts}{The IDs of state cohorts}

\item{pathToResults}{Path to temporary folder for saving objects}

\item{statePriorityVector}{A vector of all the imported states in prioritized order, ascending}

\item{excludedStates}{A vector populated with excluded states}
}
\description{
This function outputs a data.frame object which describes the movement of patients between the defined states and state duration
}
\keyword{internal}
