% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmFunctions.R
\name{kmDataPreparation}
\alias{kmDataPreparation}
\title{Prepare patientData dataframe for Kaplan-Meier plots}
\usage{
kmDataPreparation(
  patientData,
  startCohortId,
  endCohortId,
  ageInterval = NULL,
  selectedIntervals = NULL,
  survivalType = "regular"
)
}
\arguments{
\item{startCohortId}{Start cohort Id}

\item{endCohortId}{End cohort Id}

\item{ageInterval}{Indicator whether age analysis are prohibited or allowed, if 0 then prohibited}

\item{selectedIntervals}{Age intervals which were selected}

\item{survivalType}{If "nearest", then comparisons are made with the startCohortId state present nearest to the endCohortId}

\item{cohortData}{A data.frame object which is output of Cohort2Trajectory package}
}
\value{
A dataframe ready for using msm package
}
\description{
This function prepares patientData dataframe for Kaplan-Meier plots
}
\keyword{internal}
