% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runDashboard}
\alias{runDashboard}
\title{This function starts the dashboard application for comparing results from different databases}
\usage{
runDashboard(pathToResults = NULL)
}
\arguments{
\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function starts the dashboard application for comparing results from different databases
}
