% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGUI.R
\name{runGUI}
\alias{runGUI}
\title{This function initiates the connection with database and starts Shiny application}
\usage{
runGUI(
  connection,
  connectionDetails,
  pathToDriver = "./Drivers",
  dbms = "postgresql",
  cdmSchema = "ohdsi_cdm",
  cdmTmpSchema = "ohdsi_temp",
  cdmResultsSchema = "ohdsi_results",
  studyName = "MarkovAnalysis",
  pathToResults = NULL
)
}
\arguments{
\item{pathToDriver}{Path to a folder containing the JDBC driver JAR files. See downloadJdbcDrivers for instructions on how to download the relevant drivers.}

\item{dbms}{The type of DBMS running on the server. Valid values are: 'oracle','postgresql','redshift','sql server','pdw', 'netezza','bigquery','sqlite', 'sqlite extended','spark'}

\item{cdmSchema}{Schema which contains the OHDSI Common Data Model.}

\item{cdmTmpSchema}{Schema for temporary tables}

\item{cdmResultsSchema}{Schema which has the information about the cohorts created in Atlas}

\item{pathToResults}{Path to target directory where results will be saved}
}
\description{
This function initiates the connection with database and starts Shiny application
}
