\name{TrajectoryMarkovAnalysis-package}
\alias{TrajectoryMarkovAnalysis-package}
\alias{TrajectoryMarkovAnalysis}
\docType{package}
\title{
 This package creates Markov models for patient-orientated trajectories. 
}
\description{
R-package TrajectoryMarkovAnalysis specializes on modelling patient trajectories using Markov chains. The package supports discrete and continuous time Markov models, Kaplan-Meier plots, Markov trees and synthetic data generation. The package is based on OMOP CDM.
License: GPL (>= 2)
}
\details{
}
\author{
Markus Haug, haugmarkus@protonmail.com.

Maintainer: Markus Haug haugmarkus@protonmail.com
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  HealthInformaticsUT/Cohort2Trajectory,
  HealthInformaticsUT/TrajectoryViz}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
