% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisationFunctions.R
\name{visualisePatient}
\alias{visualisePatient}
\title{Visualize patient's trajectory inbetween the cohorts}
\usage{
visualisePatient(
  patientData,
  patientId,
  trajectoryStopDays = 183,
  theme = 1,
  connection,
  dbms,
  cdmTmpSchema = "ohdsi_temp",
  feature = 0
)
}
\arguments{
\item{patientData}{A Data.frame object which is resulting from}

\item{patientId}{Patient's id}

\item{trajectoryStopDays}{Nr of days which will determine the possible split between trajectories}

\item{theme}{Plot's theme indicator}

\item{connection}{Active connection}

\item{dbms}{Database management system's dialect}

\item{cdmTmpSchema}{Temp tables' schema}
}
\description{
This function outputs a patient oriented plot
}
