% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrajectoryMarkovAnalysis.R
\name{TrajectoryMarkovAnalysis}
\alias{TrajectoryMarkovAnalysis}
\title{This function outputs Markov model fitted from observed data}
\usage{
TrajectoryMarkovAnalysis(
  conn,
  dbms,
  cdmSchema,
  cdmTmpSchema,
  inputData,
  modelType,
  studyName,
  pathToResults = getwd(),
  excludedStates = c(),
  generateCostPatientTable = TRUE,
  costDomains = c("Drug", "Visit", "Procedure", "Device", "Measurement", "Observation",
    "Specimen"),
  databaseDescription = "A cool database"
)
}
\arguments{
\item{conn}{Connection to the database}

\item{dbms}{Database dialect}

\item{cdmSchema}{OHDSI OMOP CDM tabels' schema}

\item{cdmTmpSchema}{Table for creating temp tables}

\item{inputData}{Observed trajectories data outputted by Cohort2Trajectory package}

\item{modelType}{Markov model type can be "discrete" or "continuous"}

\item{studyName}{Customized study name}

\item{pathToResults}{The path where results will be saved}

\item{excludedStates}{States which have to be discarded from the study}

\item{generateCostPatientTable}{Boolean for generating cost_person table}

\item{costDomains}{Cost domains to include in cost analysis}

\item{databaseDescription}{Information about the OMOP CDM database data}
}
\description{
This function outputs Markov model fitted from observed data
}
